requirements.txt
alembic==1.0.8
atomicwrites==1.3.0
attrs==19.1.0
bidict==0.18.0
blinker==1.4
branca==0.3.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
csvvalidator==1.2
cycler==0.10.0
dateparser==0.7.1
dominate==2.3.5
dpath==1.4.2
Faker==1.0.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
folium==0.8.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lxml==4.3.2
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==7.0.0
numpy==1.16.2
pandas==0.24.1
pandasql==0.7.3
pbr==5.2.0
plaid-python==2.4.1
pluggy==0.9.0
progressbar2==3.39.2
py==1.8.0
pyparsing==2.3.1
pytest==4.4.1
python-dateutil==2.8.0
python-editor==1.0.4
python-utils==2.3.0
pytz==2018.9
PyYAML==3.13
regex==2019.3.12
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
SQLAlchemy==1.3.0
sqlparse==0.2.4
tabulate==0.8.3
text-unidecode==1.2
tzlocal==1.5.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
